oauth2-with-fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-dotenv
jinja2==3.1.2
